+++ CHANGES.md
@@ -3,7 +3,8 @@
 ## 0.9 (unreleased)
 
 
-- Nothing changed yet.
+- Added bulk import tests.
+- Added pydantic to dependencies.
 
 
 ## 0.8 (2024-04-18)